// dashboard/src/types/order.ts
export type OrderStatus = "Pending" | "Processing" | "Completed" | "Cancelled"

export type OrderItem = {
  id: string
  productId: string
  productName: string
  productImageUrl: string | null
  quantity: number
  unitPrice: number
  lineTotal: number
}

export type Order = {
  id: string
  status: OrderStatus
  totalAmount: number
  createdAt: string
  user?: {
    id: string
    name: string
    email: string
  }
  items: OrderItem[]
}